allowMethods in handleClaim
The metods Claim#allowEdit, Claim#allowBuild and Claim#allowAccess are not managed by the single method Claim#handleAction (they still exist but are deprecated). For this I added some parameters for the ClaimPermission's objects (I tried to not abuse with that but I also tried to minimize the number of identification of the ClaimPermission), I also added the ClaimPermission "Edit" which is only to fit with the mechanism of Claim#handleAction.
The methods Claim#allowBreak and Claim#allowGrantPermission are not managed in Claim#handleAction because the tests are really different.
The last one also call the event AllowPlayerClaimPermissionEvent so I still use the enum ClaimPermissionType (before it was AllowedPermissionType) because I didn't know if creating an GrantPermission object in the enum ClaimPermission could work.