Better handling of usernames with periods
Fixes #1360

If a trust/untrust target is not bracketed and contains a `.` character, this commit checks to see if the designated target is a player first, and if so applies all player logic. Otherwise, it falls back to the current logic of permission handling. This fixes Floodgate compatibility with GriefPrevention, as well as the rare real Mojang account with the `.` character in their name: https://namemc.com/profile/..Shadow...1

This commit also clarifies that permissions were the intepreted target in the trust success method.