Remove `TheUnderground` from softdepend
See rationale on issue #2009, in particular:
https://github.com/TechFortress/GriefPrevention/issues/2009#issuecomment-1436095403

... but recent changes to Paper have caused a circular dependency error
to be triggered by this - we (soft-!)depend on TheUnderground, which
depends on GriefPrevention, and boom.  I'd argue that a softdepend
should not be considered a circular dependency really, but it is,
currently.

TheUnderground appears to only work on 1.10, and current GriefPrevention
builds don't support anything that ancient, so there doesn't seem to be
much point in maintaining it in the softdepends list.

Note that it was already removed from `loadbefore` in 6d0a1b9