Fix PVP by pet inconsistencies
This is (technically) a behavioral change.
Pets are prevented from engaging in PVP attacks if the defender is PVP immune. Theoretically this state is unable to occur in vanilla, but plugins could add additional triggers for pets attacking.
Pets are no longer checked with PVP rules if PVP rules are not enabled. The assumption is that either vanilla PVP protection is in place or another PVP plugin is in place.
- In vanilla, pets will not engage in PVP unprompted. Disabling PVP disables other players causing pets to attack, so an attack must be retaliatory.
- If another PVP plugin is in place, it should handle all of PVP.